TMDM-7953: "tMDMDelete can't delete all records of one entity"
Fix possible incorrect name in delete statement (missing DB-specific name length restriction).
Don't update inbound FK to null when FK is mandatory (as NOT NULL constraint exists in this case).
Backport of commit 0a6a5e153c007f35ed13fa32bcf46edae7976bc8 for 5.4 / 5.5 branches.
(https://jira.talendforge.org/browse/TMDM-7953)

+++ org.talend.mdm.core/src/com/amalto/core/storage/hibernate/HibernateStorage.java
@@ -803,24 +803,28 @@ public class HibernateStorage implements Storage {
                             Set<ReferenceFieldMetadata> references = internalRepository.accept(new ForeignKeyIntegrity(typeToDelete));
                             // Empty values from intermediate tables to this non instantiable type and unset inbound references
                             for (ReferenceFieldMetadata reference : references) {
-                                if (reference.isMany()) {
-                                    String collectionTableName = reference.getContainingType().getName() + '_' + reference.getName() + '_' + reference.getReferencedType().getName();
-                                    String formattedTableName = MappingGenerator.formatSQLName(collectionTableName.toUpperCase(), tableResolver.getNameMaxLength());
-                                    session.createSQLQuery("delete from " + formattedTableName).executeUpdate(); //$NON-NLS-1$
-                                } else {
-                                    String referenceTableName = MappingGenerator.formatSQLName(reference.getContainingType().getName().toUpperCase(), tableResolver.getNameMaxLength());
-                                    List<String> fkColumnNames;
-                                    if (reference.getReferencedField() instanceof CompoundFieldMetadata) {
-                                        FieldMetadata[] fields = ((CompoundFieldMetadata) reference.getReferencedField()).getFields();
-                                        fkColumnNames = new ArrayList<String>(fields.length);
-                                        for (FieldMetadata field : fields) {
-                                            fkColumnNames.add(reference.getName() + "_" + field.getName());
-                                        }
+                                if (!reference.isMandatory()) {
+                                    if (reference.isMany()) {
+                                        String collectionTableName = reference.getContainingType().getName() + '_' + reference.getName() + '_' + reference.getReferencedType().getName();
+                                        String formattedTableName = MappingGenerator.formatSQLName(collectionTableName.toUpperCase(), tableResolver.getNameMaxLength());
+                                        session.createSQLQuery("delete from " + formattedTableName).executeUpdate(); //$NON-NLS-1$
                                     } else {
-                                        fkColumnNames = Collections.singletonList(reference.getName() + "_" + reference.getReferencedField().getName());
-                                    }
-                                    for (String fkColumnName : fkColumnNames) {
-                                        session.createSQLQuery("update " + referenceTableName + " set " + fkColumnName + " = NULL").executeUpdate(); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
+                                        String referenceTableName = MappingGenerator.formatSQLName(reference.getContainingType().getName().toUpperCase(), tableResolver.getNameMaxLength());
+                                        List<String> fkColumnNames;
+                                        if (reference.getReferencedField() instanceof CompoundFieldMetadata) {
+                                            FieldMetadata[] fields = ((CompoundFieldMetadata) reference.getReferencedField()).getFields();
+                                            fkColumnNames = new ArrayList<String>(fields.length);
+                                            for (FieldMetadata field : fields) {
+                                                String columnName = MappingGenerator.formatSQLName(reference.getName() + "_" + field.getName(), tableResolver.getNameMaxLength()); //$NON-NLS-1
+                                                fkColumnNames.add(columnName);
+                                            }
+                                        } else {
+                                            String columnName = MappingGenerator.formatSQLName(reference.getName() + "_" + reference.getReferencedField().getName(), tableResolver.getNameMaxLength()); //$NON-NLS-1
+                                            fkColumnNames = Collections.singletonList(columnName);
+                                        }
+                                        for (String fkColumnName : fkColumnNames) {
+                                            session.createSQLQuery("update " + referenceTableName + " set " + fkColumnName + " = NULL").executeUpdate(); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
+                                        }
                                     }
                                 }
                             }
@@ -840,7 +844,8 @@ public class HibernateStorage implements Storage {
                             }
                             // Delete the type instances
                             String className = storageClassLoader.getClassFromType(typeToDelete).getName();
-                            session.createQuery("delete from " + className).executeUpdate(); //$NON-NLS-1$
+                            String tableName = MappingGenerator.formatSQLName(className, tableResolver.getNameMaxLength());
+                            session.createQuery("delete from " + tableName).executeUpdate(); //$NON-NLS-1$
                             // Clean up full text indexes
                             if (dataSource.supportFullText()) {
                                 FullTextSession fullTextSession = Search.getFullTextSession(session);
